Source_code/requirements.txt
# Core
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0

# PyTorch — install the version matching your CUDA version from https://pytorch.org
torch>=2.0.0
torchvision>=0.15.0
